# tools/swagger.yaml
openapi: 3.0.0
info:
  title: Students API
  description: Students API Documentation with SwaggerUI.
  version: 0.1
servers:
  - url: https://students-api.azurewebsites.net
  - url: http://localhost:3000/
    description: :D
paths:
  /get-all:
    get:
      summary: Returns a list of users.
      description: Makes a request to DB and response with all users list.
      responses:
        200:    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string
        400:
          description: The specified user ID is invalid (e.g. not a number).
        404:
          description: A user with the specified ID was not found.

  /get-one/{student_id}:
    get:
      summary: Returns a single user.
      description: Makes a request to DB and response with a single user data.
      parameters:
        - in: path
          name: student_id
          required: true
          type: string
          description: Put a valid student ID, use the structure "21100001"
      responses:
        200:    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string
        400:
          description: The specified user ID is invalid (e.g. not a number).
        404:
          description: A user with the specified ID was not found.


  /add-student:
    post:
      summary: Adds a new student.
      description: Makes a request to DB and adds new student data.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:          # <!--- form field name
                  type: string
                grade:    # <!--- form field name
                  type: integer
                group:
                  type: string
                email:
                  type: string
                adress:
                  type: string
                nbhd:
                  type: string
                state:
                  type: string
              required:
                - name
                - grade
                - group
                - email
                - adress
                - nbhd
                - state
      responses:
        200:    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string
        400:
          description: The specified user ID is invalid (e.g. not a number).
        404:
          description: A user with the specified ID was not found.

        

  /update-student/{student_id}:
    put:
      summary: Update student data.
      description: Makes a request to DB and update student data.
      parameters:
        - in: path
          name: student_id
          required: true
          type: string
          description: Put a valid student ID, use the structure "21100001"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:          # <!--- form field name
                  type: string
                grade:    # <!--- form field name
                  type: integer
                group:
                  type: string
                email:
                  type: string
                adress:
                  type: string
                nbhd:
                  type: string
                state:
                  type: string
              required:
                - name
                - grade
                - group
                - email
                - adress
                - nbhd
                - state
      responses:
        200:    # status code
          description: Update is OK
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string
        400:
          description: The specified user ID is invalid.
        404:
          description: A user with the specified ID was not found.


  /delete-student/{student_id}:
    delete:
      summary: Delete a specified student.
      description: Delete students from BD.
      parameters:
        - in: path
          name: student_id
          required: true
          type: string
          description: Put a valid student ID, use the structure "21100001"
      responses:
        200:    # status code
          description: A single ID for deleted student
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string
        400:
          description: The specified user ID is invalid (e.g. not a number).
        404:
          description: A user with the specified ID was not found.